ft(update-cart): feature update cart
- implement update cart endpoint integration,
- implement update cart by clicking on add or minus

[WIP ft-update-cart-cp-90]
